# [`Optimization`]: `HexToBytes` and `TryParse` of `UInt160`, `UInt256`