--- a/benchmarks/Neo.Extensions.Benchmarks/Benchmark.StringExtensions.cs
+++ b/benchmarks/Neo.Extensions.Benchmarks/Benchmark.StringExtensions.cs
@@ -10,7 +10,6 @@
 // modifications are permitted.
 
 using BenchmarkDotNet.Attributes;
-using System.Globalization;
 
 namespace Neo.Extensions
 {