--- a/benchmarks/Neo.Extensions.Benchmarks/Program.cs
+++ b/benchmarks/Neo.Extensions.Benchmarks/Program.cs
@@ -12,5 +12,5 @@
 using BenchmarkDotNet.Running;
 using Neo.Extensions;
 
-// BenchmarkRunner.Run(typeof(Benchmark_ByteArrayComparer));
+BenchmarkRunner.Run(typeof(Benchmark_ByteArrayComparer));
 BenchmarkRunner.Run(typeof(Benchmark_StringExtensions));