# Description

Hex decode(`HexToBytes`) is used in many places.
This patch uses a new feature of .net9 to optimize it, and simplifies the implementation of `UInt160 .TryParse` and `UInt256. TryParse`.

Perf test on a simple 32-bytes case:
Before:
```
BenchmarkDotNet v0.14.0, macOS Sequoia 15.1.1 (24B2091) [Darwin 24.1.0]
Apple M4 Pro, 1 CPU, 14 logical and 14 physical cores
.NET SDK 9.0.100
  [Host]     : .NET 9.0.0 (9.0.24.52809), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.0 (9.0.24.52809), Arm64 RyuJIT AdvSIMD


| Method     | Mean     | Error   | StdDev  |
|----------- |---------:|--------:|--------:|
| HexToBytes | 196.0 ns | 0.48 ns | 0.45 ns |
```

After:
```
BenchmarkDotNet v0.14.0, macOS Sequoia 15.1.1 (24B2091) [Darwin 24.1.0]
Apple M4 Pro, 1 CPU, 14 logical and 14 physical cores
.NET SDK 9.0.100
  [Host]     : .NET 9.0.0 (9.0.24.52809), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 9.0.0 (9.0.24.52809), Arm64 RyuJIT AdvSIMD


| Method     | Mean     | Error     | StdDev    |
|----------- |---------:|----------:|----------:|
| HexToBytes | 8.502 ns | 0.0245 ns | 0.0205 ns |
```

Uses `Span.Slice` instead of `Substring` also improves performance.

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules